+++ lib/cli_commands.dart
@@ -22,20 +22,22 @@ late FlavorHelper flavorHelper;
 /// Create splash screens for Android and iOS
 void createSplash({String? path, String? flavor}) {
   if (flavor != null) {
-    print('''
+    print(
+      '''
 ╔════════════════════════════════════════════════════════════════════════════╗
 ║                              Flavor detected!                              ║
 ╠════════════════════════════════════════════════════════════════════════════╣
 ║ Setting up the $flavor flavor.
 ╚════════════════════════════════════════════════════════════════════════════╝
-''');
+''',
+    );
   }
 
   // It is important that the flavor setup occurs as soon as possible.
   // So before we generate anything, we need to setup the flavor (even if it's the default one).
   flavorHelper = FlavorHelper(flavor);
 
-  var config = getConfig(configFile: path);
+  final config = getConfig(configFile: path);
   _checkConfig(config);
   createSplashByConfig(config);
 }
@@ -44,43 +46,58 @@ void createSplash({String? path, String? flavor}) {
 void createSplashByConfig(Map<String, dynamic> config) {
   // Preparing all the data for later usage
   final String? image = _checkImageExists(config: config, parameter: 'image');
-  final String? darkImage =
+  String? darkImage =
       _checkImageExists(config: config, parameter: 'image_dark');
   final String? brandingImage =
       _checkImageExists(config: config, parameter: 'branding');
-  final String? brandingDarkImage =
+  String? brandingDarkImage =
       _checkImageExists(config: config, parameter: 'branding_dark');
   final String? color = parseColor(config['color']);
-  final String? darkColor = parseColor(config['color_dark']);
+  String? darkColor = parseColor(config['color_dark']);
   final String? backgroundImage =
       _checkImageExists(config: config, parameter: 'background_image');
-  final String? darkBackgroundImage =
+  String? darkBackgroundImage =
       _checkImageExists(config: config, parameter: 'background_image_dark');
-  final plistFiles = config['info_plist_files'];
-  String gravity = (config['fill'] ?? false) ? 'fill' : 'center';
-  if (config['android_gravity'] != null) gravity = config['android_gravity'];
-  final brandingGravity = config['branding_mode'] ?? 'bottom';
-  final bool fullscreen = config['fullscreen'] ?? false;
-  final String iosContentMode = config['ios_content_mode'] ?? 'center';
-  final webImageMode = (config['web_image_mode'] ?? 'center');
+  if (darkImage == null && darkColor == null) {
+    if (image != null) {
+      darkImage = image;
+    } else {
+      darkColor = color;
+    }
+    brandingDarkImage = brandingImage;
+    darkBackgroundImage = backgroundImage;
+  }
+
+  final plistFiles = config['info_plist_files'] as List<String>?;
+  String gravity = (config['fill'] as bool? ?? false) ? 'fill' : 'center';
+  if (config['android_gravity'] != null) {
+    gravity = config['android_gravity'] as String;
+  }
+  final brandingGravity = config['branding_mode'] as String? ?? 'bottom';
+  final bool fullscreen = config['fullscreen'] as bool? ?? false;
+  final String iosContentMode =
+      config['ios_content_mode'] as String? ?? 'center';
+  final webImageMode = config['web_image_mode'] as String? ?? 'center';
   String? android12Image;
   String? android12DarkImage;
   String? android12IconBackgroundColor;
   String? darkAndroid12IconBackgroundColor;
 
   if (config['android_12'] != null) {
-    var android12Config = config['android_12'];
+    final android12Config = config['android_12'] as Map<String, dynamic>;
     android12Image =
         _checkImageExists(config: android12Config, parameter: 'image');
     android12DarkImage =
-        _checkImageExists(config: android12Config, parameter: 'image_dark');
+        _checkImageExists(config: android12Config, parameter: 'image_dark') ??
+            android12Image;
     android12IconBackgroundColor =
         parseColor(android12Config['icon_background_color']);
     darkAndroid12IconBackgroundColor =
-        parseColor(android12Config['icon_background_color_dark']);
+        parseColor(android12Config['icon_background_color_dark']) ??
+            android12IconBackgroundColor;
   }
 
-  if (!config.containsKey('android') || config['android']) {
+  if (!config.containsKey('android') || config['android'] as bool) {
     if (Directory('android').existsSync()) {
       _createAndroidSplash(
         imagePath: image,
@@ -104,7 +121,7 @@ void createSplashByConfig(Map<String, dynamic> config) {
     }
   }
 
-  if (!config.containsKey('ios') || config['ios']) {
+  if (!config.containsKey('ios') || config['ios'] as bool) {
     if (Directory('ios').existsSync()) {
       _createiOSSplash(
         imagePath: image,
@@ -125,16 +142,17 @@ void createSplashByConfig(Map<String, dynamic> config) {
     }
   }
 
-  if (!config.containsKey('web') || config['web']) {
+  if (!config.containsKey('web') || config['web'] as bool) {
     if (Directory('web').existsSync()) {
       _createWebSplash(
-          imagePath: image,
-          darkImagePath: darkImage,
-          backgroundImage: backgroundImage,
-          darkBackgroundImage: darkBackgroundImage,
-          color: color,
-          darkColor: darkColor,
-          imageMode: webImageMode);
+        imagePath: image,
+        darkImagePath: darkImage,
+        backgroundImage: backgroundImage,
+        darkBackgroundImage: darkBackgroundImage,
+        color: color,
+        darkColor: darkColor,
+        imageMode: webImageMode,
+      );
     } else {
       print('Web folder not found, skipping web splash update...');
     }
@@ -142,8 +160,9 @@ void createSplashByConfig(Map<String, dynamic> config) {
 
   const String _greet = '''
 
-Native splash complete. 👍
+✅ Native splash complete.
 Now go finish building something awesome! 💪 You rock! 🤘🤩
+Like the package? Please give it a 👍 here: https://pub.dev/packages/flutter_native_splash
 ''';
 
   const String _whatsNew = '''
@@ -161,10 +180,10 @@ Now go finish building something awesome! 💪 You rock! 🤘🤩
 
 /// Remove any splash screen by setting the default white splash
 void removeSplash({String? path}) {
-  print('Restoring Flutter\'s default native splash screen...');
-  var config = getConfig(configFile: path);
+  print("Restoring Flutter's default native splash screen...");
+  final config = getConfig(configFile: path);
 
-  var removeConfig = <String, dynamic>{
+  final removeConfig = <String, dynamic>{
     'color': '#ffffff',
     'color_dark': '#000000'
   };
@@ -181,24 +200,28 @@ void removeSplash({String? path}) {
     removeConfig['web'] = config['web'];
   }
 
+  /// Checks if the image that was specified in the config file does exist.
+  /// If not the developer will receive an error message and the process will exit.
+  if (config.containsKey('info_plist_files')) {
+    removeConfig['info_plist_files'] = config['info_plist_files'];
+  }
   createSplashByConfig(removeConfig);
 }
 
-/// Checks if the image that was specified in the config file does exist.
-/// If not the developer will receive an error message and the process will exit.
 String? _checkImageExists({
   required Map<String, dynamic> config,
   required String parameter,
 }) {
-  String image = config[parameter] ?? '';
+  final String image = config[parameter].toString();
   if (image.isNotEmpty && !File(image).existsSync()) {
     print('The file "$image" set as the parameter "$parameter" was not found.');
     exit(1);
   }
 
   if (image.isNotEmpty && p.extension(image).toLowerCase() != '.png') {
-    print('Unsupported file format: $image'
-        '  Your image must be a PNG file.');
+    print(
+      'Unsupported file format: $image  Your image must be a PNG file.',
+    );
     exit(1);
   }
   return image == '' ? null : image;
@@ -223,37 +246,34 @@ Map<String, dynamic> getConfig({String? configFile}) {
     filePath = 'pubspec.yaml';
   }
 
-  if (!File(filePath).existsSync()) {
-    print('The config file `$filePath` was not found.');
-    exit(1);
-  }
-
-  final Map yamlMap = loadYaml(File(filePath).readAsStringSync());
+  final Map yamlMap = loadYaml(File(filePath).readAsStringSync()) as Map;
 
   if (yamlMap['flutter_native_splash'] is! Map) {
-    throw Exception('Your `$filePath` file does not contain a '
-        '`flutter_native_splash` section.');
+    throw Exception(
+      'Your `$filePath` file does not contain a '
+      '`flutter_native_splash` section.',
+    );
   }
 
   // yamlMap has the type YamlMap, which has several unwanted side effects
-  return _yamlToMap(yamlMap['flutter_native_splash']);
+  return _yamlToMap(yamlMap['flutter_native_splash'] as YamlMap);
 }
 
 Map<String, dynamic> _yamlToMap(YamlMap yamlMap) {
-  Map<String, dynamic> map = <String, dynamic>{};
-  for (MapEntry<dynamic, dynamic> entry in yamlMap.entries) {
+  final Map<String, dynamic> map = <String, dynamic>{};
+  for (final MapEntry<dynamic, dynamic> entry in yamlMap.entries) {
     if (entry.value is YamlList) {
-      var list = <String>[];
-      for (var value in (entry.value as YamlList)) {
+      final list = <String>[];
+      for (final value in entry.value as YamlList) {
         if (value is String) {
           list.add(value);
         }
       }
-      map[entry.key] = list;
+      map[entry.key as String] = list;
     } else if (entry.value is YamlMap) {
-      map[entry.key] = _yamlToMap(entry.value);
+      map[entry.key as String] = _yamlToMap(entry.value as YamlMap);
     } else {
-      map[entry.key] = entry.value;
+      map[entry.key as String] = entry.value;
     }
   }
   return map;
@@ -263,62 +283,83 @@ Map<String, dynamic> _yamlToMap(YamlMap yamlMap) {
 /// If they do, the developer will get a message where the issue is.
 void _checkConfig(Map<String, dynamic> config) {
   if (config.containsKey('color') && config.containsKey('background_image')) {
-    print('Your `flutter_native_splash` section cannot not contain both a '
-        '`color` and `background_image`.');
+    print(
+      'Your `flutter_native_splash` section cannot not contain both a '
+      '`color` and `background_image`.',
+    );
     exit(1);
   }
 
   if (!config.containsKey('color') && !config.containsKey('background_image')) {
-    print('Your `flutter_native_splash` section does not contain a `color` or '
-        '`background_image`.');
+    print(
+      'Your `flutter_native_splash` section does not contain a `color` or '
+      '`background_image`.',
+    );
     exit(1);
   }
 
   if (config.containsKey('color_dark') &&
       config.containsKey('background_image_dark')) {
-    print('Your `flutter_native_splash` section cannot not contain both a '
-        '`color_dark` and `background_image_dark`.');
+    print(
+      'Your `flutter_native_splash` section cannot not contain both a '
+      '`color_dark` and `background_image_dark`.',
+    );
     exit(1);
   }
 
   if (config.containsKey('image_dark') &&
       !config.containsKey('color_dark') &&
       !config.containsKey('background_image_dark')) {
-    print('Your `flutter_native_splash` section contains `image_dark` but '
-        'does not contain a `color_dark` or a `background_image_dark`.');
+    print(
+      'Your `flutter_native_splash` section contains `image_dark` but '
+      'does not contain a `color_dark` or a `background_image_dark`.',
+    );
+    exit(1);
+  }
+
+  if (config.containsKey('branding_dark') && !config.containsKey('branding')) {
+    print(
+      'Your `flutter_native_splash` section contains `branding_dark` but '
+      'does not contain a `branding`.',
+    );
     exit(1);
   }
 
   if (config.containsKey('android_12') && config['android_12'] is Map) {
-    Map android12Config = config['android_12'];
+    final android12Config = config['android_12'] as Map;
     if (android12Config.containsKey('image_dark') &&
         !config.containsKey('color_dark') &&
         !config.containsKey('background_image_dark')) {
-      print('Your `flutter_native_splash` section contains '
-          '`android_12:image_dark` but does not contain a `color_dark` or a '
-          '`background_image_dark`.');
+      print(
+        'Your `flutter_native_splash` section contains '
+        '`android_12:image_dark` but does not contain a `color_dark` or a '
+        '`background_image_dark`.',
+      );
       exit(1);
     }
     if (android12Config.containsKey('image_dark') &&
         !config.containsKey('color_dark') &&
         !config.containsKey('icon_background_color_dark')) {
-      print('Your `flutter_native_splash` section contains '
-          '`android_12:icon_background_color_dark` but does not contain a '
-          '`color_dark` or a `background_image_dark`.');
+      print(
+        'Your `flutter_native_splash` section contains '
+        '`android_12:icon_background_color_dark` but does not contain a '
+        '`color_dark` or a `background_image_dark`.',
+      );
       exit(1);
     }
   }
 }
 
 @visibleForTesting
-String? parseColor(var color) {
-  if (color is int) color = color.toString().padLeft(6, '0');
+String? parseColor(dynamic color) {
+  dynamic colorValue = color;
+  if (colorValue is int) colorValue = colorValue.toString().padLeft(6, '0');
 
-  if (color is String) {
-    color = color.replaceAll('#', '').replaceAll(' ', '');
-    if (color.length == 6) return color;
+  if (colorValue is String) {
+    colorValue = colorValue.replaceAll('#', '').replaceAll(' ', '');
+    if (colorValue.length == 6) return colorValue;
   }
-  if (color == null) return null;
+  if (colorValue == null) return null;
 
   throw Exception('Invalid color value');
 }