+++ lib/web.dart
@@ -18,31 +18,39 @@ void _createWebSplash({
   required String? darkBackgroundImage,
 }) {
   if (!File(_webIndex).existsSync()) {
-    print('[Web] ' + _webIndex + ' not found.  Skipping Web.');
+    print('[Web] $_webIndex not found.  Skipping Web.');
     return;
   }
 
   darkImagePath ??= imagePath;
-  createWebImages(imagePath: imagePath, webSplashImages: [
-    _WebLaunchImageTemplate(fileName: 'light-1x.png', pixelDensity: 1),
-    _WebLaunchImageTemplate(fileName: 'light-2x.png', pixelDensity: 2),
-    _WebLaunchImageTemplate(fileName: 'light-3x.png', pixelDensity: 3),
-    _WebLaunchImageTemplate(fileName: 'light-4x.png', pixelDensity: 4),
-  ]);
-  createWebImages(imagePath: darkImagePath, webSplashImages: [
-    _WebLaunchImageTemplate(fileName: 'dark-1x.png', pixelDensity: 1),
-    _WebLaunchImageTemplate(fileName: 'dark-2x.png', pixelDensity: 2),
-    _WebLaunchImageTemplate(fileName: 'dark-3x.png', pixelDensity: 3),
-    _WebLaunchImageTemplate(fileName: 'dark-4x.png', pixelDensity: 4),
-  ]);
+  createWebImages(
+    imagePath: imagePath,
+    webSplashImages: [
+      _WebLaunchImageTemplate(fileName: 'light-1x.png', pixelDensity: 1),
+      _WebLaunchImageTemplate(fileName: 'light-2x.png', pixelDensity: 2),
+      _WebLaunchImageTemplate(fileName: 'light-3x.png', pixelDensity: 3),
+      _WebLaunchImageTemplate(fileName: 'light-4x.png', pixelDensity: 4),
+    ],
+  );
+  createWebImages(
+    imagePath: darkImagePath,
+    webSplashImages: [
+      _WebLaunchImageTemplate(fileName: 'dark-1x.png', pixelDensity: 1),
+      _WebLaunchImageTemplate(fileName: 'dark-2x.png', pixelDensity: 2),
+      _WebLaunchImageTemplate(fileName: 'dark-3x.png', pixelDensity: 3),
+      _WebLaunchImageTemplate(fileName: 'dark-4x.png', pixelDensity: 4),
+    ],
+  );
   createBackgroundImages(
-      backgroundImage: backgroundImage,
-      darkBackgroundImage: darkBackgroundImage);
+    backgroundImage: backgroundImage,
+    darkBackgroundImage: darkBackgroundImage,
+  );
   _createSplashCss(
-      color: color,
-      darkColor: darkColor,
-      darkBackgroundImage: darkBackgroundImage,
-      backgroundImage: backgroundImage);
+    color: color,
+    darkColor: darkColor,
+    darkBackgroundImage: darkBackgroundImage,
+    backgroundImage: backgroundImage,
+  );
   _createSplashJs();
   updateIndex(imageMode: imageMode, imagePath: imagePath);
 }
@@ -51,7 +59,7 @@ void createBackgroundImages({
   required String? backgroundImage,
   required String? darkBackgroundImage,
 }) {
-  const backgroundDestination = _webSplashImagesFolder + 'light-background.png';
+  const backgroundDestination = '${_webSplashImagesFolder}light-background.png';
   if (backgroundImage == null) {
     final file = File(backgroundDestination);
     if (file.existsSync()) file.deleteSync();
@@ -63,7 +71,7 @@ void createBackgroundImages({
   }
 
   const darkBackgroundDestination =
-      _webSplashImagesFolder + 'dark-background.png';
+      '${_webSplashImagesFolder}dark-background.png';
   if (darkBackgroundImage == null) {
     final file = File(darkBackgroundDestination);
     if (file.existsSync()) file.deleteSync();
@@ -75,74 +83,82 @@ void createBackgroundImages({
   }
 }
 
-void createWebImages(
-    {required String? imagePath,
-    required List<_WebLaunchImageTemplate> webSplashImages}) {
+void createWebImages({
+  required String? imagePath,
+  required List<_WebLaunchImageTemplate> webSplashImages,
+}) {
   if (imagePath == null) {
-    for (var template in webSplashImages) {
+    for (final template in webSplashImages) {
       final file = File(_webSplashImagesFolder + template.fileName);
       if (file.existsSync()) file.deleteSync();
     }
   } else {
     final image = decodeImage(File(imagePath).readAsBytesSync());
     if (image == null) {
-      print(imagePath + ' could not be read');
+      print('$imagePath could not be read');
       exit(1);
     }
     print('[Web] Creating images');
-    for (var template in webSplashImages) {
+    for (final template in webSplashImages) {
       _saveImageWeb(template: template, image: image);
     }
   }
 }
 
-void _saveImageWeb(
-    {required _WebLaunchImageTemplate template, required Image image}) {
-  var newFile = copyResize(
+void _saveImageWeb({
+  required _WebLaunchImageTemplate template,
+  required Image image,
+}) {
+  final newFile = copyResize(
     image,
     width: image.width * template.pixelDensity ~/ 4,
     height: image.height * template.pixelDensity ~/ 4,
     interpolation: Interpolation.linear,
   );
 
-  var file = File(_webSplashImagesFolder + template.fileName);
+  final file = File(_webSplashImagesFolder + template.fileName);
   file.createSync(recursive: true);
   file.writeAsBytesSync(encodePng(newFile));
 }
 
-void _createSplashCss(
-    {required String? color,
-    required String? darkColor,
-    required String? backgroundImage,
-    required String? darkBackgroundImage}) {
+void _createSplashCss({
+  required String? color,
+  required String? darkColor,
+  required String? backgroundImage,
+  required String? darkBackgroundImage,
+}) {
   print('[Web] Creating CSS');
   color ??= '000000';
   darkColor ??= color;
   var cssContent = _webCss
-      .replaceFirst('[LIGHTBACKGROUNDCOLOR]', '#' + color)
-      .replaceFirst('[DARKBACKGROUNDCOLOR]', '#' + darkColor);
+      .replaceFirst('[LIGHTBACKGROUNDCOLOR]', '#$color')
+      .replaceFirst('[DARKBACKGROUNDCOLOR]', '#$darkColor');
 
   if (backgroundImage == null) {
     cssContent = cssContent.replaceFirst('[LIGHTBACKGROUNDIMAGE]', '');
   } else {
-    cssContent = cssContent.replaceFirst('[LIGHTBACKGROUNDIMAGE]',
-        'background-image: url("img/light-background.png");');
+    cssContent = cssContent.replaceFirst(
+      '[LIGHTBACKGROUNDIMAGE]',
+      'background-image: url("img/light-background.png");',
+    );
   }
 
   if (backgroundImage == null) {
     cssContent = cssContent.replaceFirst('[DARKBACKGROUNDIMAGE]', '');
   } else {
-    cssContent = cssContent.replaceFirst('[DARKBACKGROUNDIMAGE]',
-        'background-image: url("img/dark-background.png");');
+    cssContent = cssContent.replaceFirst(
+      '[DARKBACKGROUNDIMAGE]',
+      'background-image: url("img/dark-background.png");',
+    );
   }
 
-  var file = File(_webFolder + _webRelativeStyleFile);
+  final file = File(_webFolder + _webRelativeStyleFile);
   file.createSync(recursive: true);
   file.writeAsStringSync(cssContent);
 }
 
 void _createSplashJs() {
-  var file = File(_webFolder + _webRelativeJSFile);
+  final file = File(_webFolder + _webRelativeJSFile);
   file.createSync(recursive: true);
   file.writeAsStringSync(_webJS);
 }
@@ -150,7 +166,7 @@ void _createSplashJs() {
 void updateIndex({required String imageMode, required String? imagePath}) {
   print('[Web] Updating index.html');
   final webIndex = File(_webIndex);
-  var lines = webIndex.readAsLinesSync();
+  final lines = webIndex.readAsLinesSync();
 
   var foundExistingStyleSheet = false;
   bool foundExistingMetaViewport = false;
@@ -165,7 +181,7 @@ void updateIndex({required String imageMode, required String? imagePath}) {
       '<meta content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" name="viewport"/>';
   const jsLink = '<script src="splash/splash.js"></script>';
   for (var x = 0; x < lines.length; x++) {
-    var line = lines[x];
+    final line = lines[x];
 
     if (line.contains(styleSheetLink)) {
       foundExistingStyleSheet = true;
@@ -187,25 +203,21 @@ void updateIndex({required String imageMode, required String? imagePath}) {
   }
 
   if (!foundExistingStyleSheet) {
-    lines[headCloseTagLine] =
-        '  ' + styleSheetLink + '\n' + lines[headCloseTagLine];
+    lines[headCloseTagLine] = '  $styleSheetLink\n${lines[headCloseTagLine]}';
   }
   if (!foundExistingMetaViewport) {
-    lines[headCloseTagLine] =
-        '  ' + metaViewport + '\n' + lines[headCloseTagLine];
+    lines[headCloseTagLine] = '  $metaViewport\n${lines[headCloseTagLine]}';
   }
 
   if (!foundExistingJs) {
-    lines[headCloseTagLine] = '  ' + jsLink + '\n' + lines[headCloseTagLine];
+    lines[headCloseTagLine] = '  $jsLink\n${lines[headCloseTagLine]}';
   }
 
   if (existingPictureLine == 0) {
     if (imagePath != null) {
       for (var x = _indexHtmlPicture.length - 1; x >= 0; x--) {
         lines[bodyOpenTagLine + 1] =
-            _indexHtmlPicture[x].replaceFirst('[IMAGEMODE]', imageMode) +
-                '\n' +
-                lines[bodyOpenTagLine + 1];
+            '${_indexHtmlPicture[x].replaceFirst('[IMAGEMODE]', imageMode)}\n${lines[bodyOpenTagLine + 1]}';
       }
     }
   } else {
@@ -216,7 +228,9 @@ void updateIndex({required String imageMode, required String? imagePath}) {
       }
     } else {
       lines.removeRange(
-          existingPictureLine, existingPictureLine + _indexHtmlPicture.length);
+        existingPictureLine,
+        existingPictureLine + _indexHtmlPicture.length,
+      );
     }
   }
   webIndex.writeAsStringSync(lines.join('\n'));